simple routing between todo page and history page
