+++ src/History.vue
@@ -1,0 +1,18 @@
+<script setup>
+import { ref, computed } from "vue";
+</script>
+
+<template>
+  <div class="container mx-auto p-4 max-w-screen-sm min-h-screen">
+    <header class="my-8">
+      <h1 class="font-semibold text-2xl text-pink-500">History</h1>
+    </header>
+
+    <main>
+      <a href="#/">go back</a>
+    </main>
+  </div>
+</template>
+
+<style scoped>
+</style>
